Split the point module into event geography and control points
